Reorganization, adding functions for dbpm
