+++ backend/handlers/math/market/calculateuserownedshares.draft.go.txt
@@ -29,7 +29,7 @@ func CalculateUserSharesOwnedOnMarket(db *gorm.DB, username string, marketID uin
 	}
 
 	// get a timeline of probability changes for the market
-	probabilityChanges := CalculateMarketProbabilities(market, allBetsOnMarket)
+	probabilityChanges := CalculateMarketProbabilitiesWPAM(market, allBetsOnMarket)
 
 	//
 	for _, bet := range userBets {
