+++ backend/handlers/math/market/marketvolume.go
@@ -3,11 +3,13 @@ package marketmath
 import "socialpredict/models"
 
 // getMarketVolume returns the total volume of trades for a given market
-func GetMarketVolume(bets []models.Bet) float64 {
+func GetMarketVolume(bets []models.Bet) uint {
 	var totalVolume float64
 	for _, bet := range bets {
 		totalVolume += bet.Amount
 	}
 
-	return totalVolume
+	totalVolumeUint := uint(totalVolume)
+
+	return totalVolumeUint
 }
